requirements.txt
requests
bs4
click
pytest
psycopg2-binary
Flask-Script
Flask-SQLAlchemy
alembic>=1.7.6
Flask-Migrate
flask
spacy
dotenv
en_core_web_sm
